User BE class

+++ src/timemanager/bll/TimeSaver.java
@@ -50,7 +50,7 @@ public class TimeSaver
         
         stopTime();
         
-        System.out.println("Time spent: " + spentTime());
+        System.out.println("Time spent: " + (int)spentTime());
     }
 
 }
